+++ aea/cli/run.py
@@ -28,10 +28,12 @@ from pathlib import Path
 from typing import cast
 
 import click
+from click import pass_context
 
 from aea.aea import AEA
 from aea.cli.common import Context, pass_ctx, logger, _try_to_load_agent_config, _try_to_load_protocols, \
     AEAConfigException
+from aea.cli.install import install
 from aea.connections.base import Connection
 from aea.crypto.base import Crypto
 from aea.crypto.helpers import _try_validate_private_key_pem_path, _create_temporary_private_key_pem_path
@@ -80,9 +82,12 @@ def _setup_connection(connection_name: str, public_key: str, ctx: Context) -> Co
 @click.command()
 @click.option('--connection', 'connection_name', metavar="CONN_NAME", type=str, required=False, default=None,
               help="The connection name. Must be declared in the agent's configuration file.")
-@pass_ctx
-def run(ctx: Context, connection_name: str):
+@click.option('--install-deps', 'install_deps', is_flag=True, required=False, default=False,
+              help="Install all the dependencies before running the agent.")
+@pass_context
+def run(click_context, connection_name: str, install_deps: bool):
     """Run the agent."""
+    ctx = cast(Context, click_context.obj)
     _try_to_load_agent_config(ctx)
     agent_name = cast(str, ctx.agent_config.agent_name)
     private_key_pem_path = cast(str, ctx.agent_config.private_key_pem_path)
@@ -100,6 +105,12 @@ def run(ctx: Context, connection_name: str):
         logger.error(str(e))
         exit(-1)
         return
+
+    if install_deps:
+        if Path("requirements.txt").exists():
+            click_context.invoke(install, requirement="requirements.txt")
+        else:
+            click_context.invoke(install)
 
     mailbox = MailBox(connection)
     agent = AEA(agent_name, mailbox, private_key_pem_path=private_key_pem_path, directory=str(Path(".")))
